deerflow/requirements_deerflow.txt
langchain>=1.2.3
langchain-openai>=1.1.7
langgraph>=1.0.6
langgraph-runtime-inmem>=0.22.1
pydantic>=2.12.5
python-dotenv>=0.9.9
